Virtual channels support removed (will be rewritten). Added progress bar.

+++ src/GlobalStats.cpp
@@ -229,7 +229,7 @@ void GlobalStats::showStats(std::ostream& out, bool detailed)
 	out << "% Total received packets: " << getReceivedPackets() << endl;
 	out << "% Total received flits: " << getReceivedFlits() << endl;
 	out << "% Received/Ideal flits Ratio: " << getReceivedIdealFlitRatio() << endl;
-	out << "% Last time flit received: " << getLastReceivedFlitTime() << '\n';
+	out << "% Last time flit received: " << (int32_t)getLastReceivedFlitTime() << '\n';
 	out << "% Global average delay (cycles): " << getAverageDelay() << endl;
 	out << "% Max delay (cycles): " << getMaxDelay() << endl;
 	out << "% Network throughput (flits/cycle): " << getAggregatedThroughput() << endl;
